requirements.txt
Django==1.9.1
django-bootstrap3==6.2.2
django-pure-pagination==0.3.0
wheel==0.24.0
wsgiref==0.1.2
